refactored PlayerWindow and its SplitPane handling
