+++ vassal-app/src/main/java/VASSAL/chat/peer2peer/DirectPeerPool.java
@@ -152,7 +152,7 @@ public class DirectPeerPool implements PeerPool, ChatControlsInitializer {
 
     Frame owner = null;
     if (GameModule.getGameModule() != null) {
-      owner = GameModule.getGameModule().getFrame();
+      owner = GameModule.getGameModule().getPlayerWindow();
     }
     frame = new JDialog(owner,Resources.getString("Peer2Peer.direct_connection")); //$NON-NLS-1$
     frame.setDefaultCloseOperation(WindowConstants.HIDE_ON_CLOSE);
