+++ vassal-app/src/main/java/VASSAL/command/AlertCommand.java
@@ -34,7 +34,7 @@ public class AlertCommand extends Command {
     Runnable runnable = new Runnable() {
       @Override
       public void run() {
-        JOptionPane.showMessageDialog(GameModule.getGameModule() == null ? null : GameModule.getGameModule().getFrame(), msg);
+        JOptionPane.showMessageDialog(GameModule.getGameModule() == null ? null : GameModule.getGameModule().getPlayerWindow(), msg);
       }
     };
     SwingUtilities.invokeLater(runnable);
