+++ vassal-app/src/main/java/VASSAL/launch/CreateModuleAction.java
@@ -38,7 +38,7 @@ public class CreateModuleAction extends GameModuleAction {
   @Override
   public void performAction(ActionEvent e) throws IOException {
     GameModule.init(new BasicModule(new ArchiveWriter((String) null)));
-    JFrame frame = GameModule.getGameModule().getFrame();
+    JFrame frame = GameModule.getGameModule().getPlayerWindow();
     frame.setVisible(true);
     ModuleEditorWindow w = new ModuleEditorWindow(GameModule.getGameModule());
     w.setLocation(0, frame.getY() + frame.getHeight());
