+++ vassal-app/src/main/java/VASSAL/launch/EditExtensionAction.java
@@ -54,7 +54,7 @@ public class EditExtensionAction extends LoadModuleAction {
     final ModuleExtension ext =
       new ModuleExtension(new ArchiveWriter(new ZipArchive(f)));
     ext.build();
-    final JFrame frame = GameModule.getGameModule().getFrame();
+    final JFrame frame = GameModule.getGameModule().getPlayerWindow();
     final ExtensionEditorWindow w =
       new ExtensionEditorWindow(GameModule.getGameModule(), ext);
     w.setLocation(0, frame.getY() + frame.getHeight());
