+++ vassal-app/src/main/java/VASSAL/launch/ModuleManagerWindow.java
@@ -121,6 +121,7 @@ import VASSAL.tools.menu.MenuItemProxy;
 import VASSAL.tools.menu.MenuManager;
 import VASSAL.tools.menu.MenuProxy;
 import VASSAL.tools.swing.Dialogs;
+import VASSAL.tools.swing.SplitPane;
 import VASSAL.tools.swing.SwingUtils;
 import VASSAL.tools.version.UpdateCheckAction;
 
@@ -151,7 +152,7 @@ public class ModuleManagerWindow extends JFrame {
 
   private CardLayout modulePanelLayout;
   private JPanel moduleView;
-  private ComponentSplitter.SplitPane serverStatusView;
+  private SplitPane serverStatusView;
 
   private MyTreeNode rootNode;
   private MyTree tree;
@@ -420,10 +421,10 @@ public class ModuleManagerWindow extends JFrame {
     serverStatusControls.setBorder(
       new TitledBorder(Resources.getString("Chat.server_status")));
 
-    serverStatusView = ComponentSplitter.split(
+    serverStatusView = ComponentSplitter.splitAsNewSplitPane(
       moduleControls,
       serverStatusControls,
-      ComponentSplitter.SplitPane.HIDE_RIGHT,
+      SplitPane.HIDE_RIGHT,
       false
     );
     serverStatusView.revalidate();
