+++ vassal-app/src/main/java/VASSAL/tools/ErrorDialog.java
@@ -86,7 +86,7 @@ public class ErrorDialog {
       logger.error("", thrown);
 
       final Frame frame = GameModule.getGameModule() == null
-        ? null : GameModule.getGameModule().getFrame();
+        ? null : GameModule.getGameModule().getPlayerWindow();
 
       SwingUtilities.invokeLater(new Runnable() {
         @Override
